Pull FancyTransition into its own file.
